+++ contracts/handlers/curve/ICurveHandler.sol
@@ -1,10 +1,6 @@
 pragma solidity ^0.5.0;
 
-interface ICurveSwap {
-    function coins(int128 arg0) external view returns (address);
-
-    function underlying_coins(int128 arg0) external view returns (address);
-
+interface ICurveHandler {
     function get_dy(
         int128 i,
         int128 j,
@@ -31,16 +27,20 @@ interface ICurveSwap {
         uint256 min_dy
     ) external;
 
-    // ren pool
+    // Curve contract only support static array
     function add_liquidity(uint256[2] calldata amounts, uint256 min_mint_amount)
         external;
 
-    // sbtc pool
     function add_liquidity(uint256[3] calldata amounts, uint256 min_mint_amount)
         external;
 
-    // susd pool
     function add_liquidity(uint256[4] calldata amounts, uint256 min_mint_amount)
+        external;
+
+    function add_liquidity(uint256[5] calldata amounts, uint256 min_mint_amount)
+        external;
+
+    function add_liquidity(uint256[6] calldata amounts, uint256 min_mint_amount)
         external;
 
     function calc_token_amount(uint256[2] calldata amounts, bool deposit)
@@ -58,11 +58,27 @@ interface ICurveSwap {
         view
         returns (uint256);
 
-    // Curve ren and sbtc pools
+    function calc_token_amount(uint256[5] calldata amounts, bool deposit)
+        external
+        view
+        returns (uint256);
+
+    function calc_token_amount(uint256[6] calldata amounts, bool deposit)
+        external
+        view
+        returns (uint256);
+
     function remove_liquidity_one_coin(
         uint256 _token_amount,
         int128 i,
         uint256 min_amount
+    ) external;
+
+    function remove_liquidity_one_coin(
+        uint256 _token_amount,
+        int128 i,
+        uint256 min_uamount,
+        bool donate_dust
     ) external;
 
     function calc_withdraw_one_coin(uint256 _token_amount, int128 i)